+++ src/main/java/com/onlineschool/schoolmanagementsystem/service/AdminCommonService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 @Service
 public interface AdminCommonService<T> {
 	public T register(T dtoObject);
-	public T fetchDetails(Long dtoObject);//get
+	public T fetchDetails(Integer dtoObject);//get
 	public T updateName(T dtoObject);//patch
 	public T updateAddress(T dtoObject);//patch
 	/*status types
